extract hardware part of microsecond_timer
