+++ firmware/hw_layer/ports/cypress/mpu_util.cpp
@@ -257,9 +257,4 @@ uintptr_t getFlashAddrSecondCopy() {
 	return FLASH_ADDR_SECOND_COPY;
 }
 
-uint32_t getTimeNowLowerNt() {
-	return port_rt_get_counter_value();
-}
-
 #endif /* EFI_PROD_CODE */
-
