+++ firmware/hw_layer/ports/stm32/stm32_common.cpp
@@ -325,7 +325,3 @@ bool isValidSerialRxPin(brain_pin_e pin) {
 }
 
 #endif /*EFI_AUX_SERIAL*/
-
-uint32_t getTimeNowLowerNt() {
-	return port_rt_get_counter_value();
-}
