+++ firmware/hw_layer/ports/stm32/stm32_common.mk
@@ -5,7 +5,8 @@ HW_LAYER_EMS += $(PROJECT_DIR)/hw_layer/ports/stm32/serial_over_usb/usbcfg.c \
 HW_LAYER_EMS_CPP += \
 	$(PROJECT_DIR)/hw_layer/ports/stm32/stm32_pins.cpp \
 	$(PROJECT_DIR)/hw_layer/ports/stm32/stm32_common.cpp \
-	$(PROJECT_DIR)/hw_layer/ports/stm32/backup_ram.cpp
+	$(PROJECT_DIR)/hw_layer/ports/stm32/backup_ram.cpp \
+	$(PROJECT_DIR)/hw_layer/microsecond_timer/microsecond_timer_gpt.cpp \
 
 RUSEFIASM = $(PROJECT_DIR)/hw_layer/ports/stm32/rusEfiStartup.S
